Fixed critical issue in exception frame state merging (prevented jython from running correctly). Added intrinsification for float/double to int/long conversions and for currentThread.

+++ dataset/GitHub_Java/oracle.graal/graal/com.oracle.max.graal.compiler/src/com/oracle/max/graal/compiler/graph/BlockMap.java
@@ -523,7 +523,7 @@ public final class BlockMap {
             }
 
             assert block.loops == 0;
-            block.loops = (long) 1 << (long) nextLoop;
+            block.loops = 1 << nextLoop;
             nextLoop++;
         }
         assert Long.bitCount(block.loops) == 1;
